Update rights basis layout
Updates the rights basis objects to extend a common `rights_basis_information` object. This object includes common data you'd find in a rights basis.

It moves the `rights_basis` field to the common objects.

Removed the minimumItems property on rights_granted but still requires the field allowing for an empty rights granted array, but still requiring conformance to the rights granted object schema if there is anything in the array.